Update README.md Correction of my name being dropped
Changed the word "dave" as in dave, to have which it was supposed to be. 

I got a notification that jarvis found yet another developer using my name without permission, Im just here to fix that mistake no drama. And thanks for this repository, Jarvis is happy to be used by a generic developer <3